Update boot.py to use modified boot descriptor
Even with extra endpoints disabled the macropad was not working with the
KVM switch. A modified boot descriptor is required for the keyboard to
work as a boot device, this was taken from the following GitHub issue:

https://github.com/adafruit/circuitpython/issues/1136#issuecomment-1002833056

An new keyboard object is created that uses the updated boot descriptor.
When booting in maintenance_mode the standard keyboard object is used.